+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- Fixed the parsing of CodeChef problems on /submit/ URLs
+
 ## [2.35.1](https://github.com/jmerle/competitive-companion/releases/tag/2.35.1) (2022-07-29)
 - Fixed the parsing of CodeChef problems
 - Updated the CodeChef parsers so that they all consistently use the shortened contest code in the group property